web almost done, city page and user page incoming

+++ app/Http/Controllers/ReservasiController.php
@@ -5,33 +5,46 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Reservasi;
 use App\ReservasiDetail;
+use App\City;
 use App\Jadwal;
 use App\Studio;
 use App\Room;
-use \Exception;
+use App\User;
+use Exception;
+use Session;
+use Redirect;
 
 class ReservasiController extends Controller
 {
   function store(Request $request){
+      $output=(object)array();
     // try {
       $user_id = $request->input('user_id');
       $room_id = $request->input('room_id');
+      $studio_id = Room::where('room_id',$room_id)->first(['studio_id'])->toArray()['studio_id'];
       $harga = Room::where('room_id','=',$room_id)->get()->toArray()[0]['room_harga'];
       $nomor_booking = strtoupper('S'.$this->getRandomHex(3));
       $nama_band = $request->input('nama_band');
       $waktu_booking = date('H:i:s');
       $tanggal = $request->input('tanggal');
       $jadwal = $request->input('jadwal');
+      if($jadwal == ''){
+        $output->code = "-1";
+        $output->status = "Belum memilih jadwal";
+        return json_encode($output);
+      }
       $list_jadwal = explode(',',$jadwal);
       $tagihan = sizeof($list_jadwal) * (int)$harga;
 
       $input = new Reservasi();
       $input->user_id =$user_id;
+      $input->studio_id =$studio_id;
       $input->room_id =$room_id;
       $input->reservasi_nomor_booking =$nomor_booking;
       $input->reservasi_nama_band =$nama_band;
       $input->reservasi_tagihan =$tagihan;
       $input->reservasi_status = '0';
+      $input->reservasi_refund = '0';
       $input->reservasi_waktu_booking = $waktu_booking;
       $input->reservasi_tanggal = $tanggal;
       try {
@@ -39,17 +52,14 @@ class ReservasiController extends Controller
         foreach ($list_jadwal as $l) {
           $detail = new ReservasiDetail();
           $detail->reservasi_id = $input->id;
-          $detail->jadwal_id = $l;
+          $detail->jadwal_id = (int)$l;
           $detail->save();
           unset($detail);
         }
-        $output=(object)array();
         $output->code = "1";
         $output->status = "Reservasi berhasil";
         $output->reservasi = $input;
       } catch (Exception $e) {
-        return $e;
-        $output=(object)array();
         $output->code = "0";
         $output->status = "Reservasi gagal";
       }
@@ -59,56 +69,64 @@ class ReservasiController extends Controller
   function edit(Request $request){
     $reservasi_id = $request->input('reservasi_id');
     $reservasi = Reservasi::where('reservasi_id','=',$reservasi_id)->get();
+    $ouput=array();
     if($reservasi->count()==0){
-      return 'Data Reservasi tidak ada';
+      $output['code']='-1';
+      $output['message'] = 'Data Reservasi tidak ada';
     } else {
-      $user_id = $request->input('user_id');
-      $studio_id = $request->input('studio_id');
-      $harga = Studio::where('studio_id','=',$studio_id)->get()->toArray()[0]['studio_harga'];
+      $reservasi->toArray();
+      $jumlah = ReservasiDetail::where('reservasi_id',$reservasi_id)->get()->count();
       $nama_band = $request->input('nama_band');
-      $waktu_booking = date('H:i:s');
-      $tanggal = $request->input('tanggal');
+      if($nama_band == '') $nama_band=$reservasi['reservasi_nama_band'];
       $jadwal = $request->input('jadwal');
       $list_jadwal = explode(',',$jadwal);
-      $tagihan = sizeof($list_jadwal) * (int)$harga;
+      if(count($list_jadwal) != $jumlah){
+        $output['code']='0';
+        $output['message'] = 'Edit reservasi gagal';
+      } else {
+        $data = [
+          'reservasi_nama_band' =>$nama_band,
+        ];
 
-      $data = [
-        'reservasi_nama_band' =>$nama_band,
-        'reservasi_tagihan'   =>$tagihan,
-        'reservasi_tanggal'   =>$tanggal
-      ];
-
-      Reservasi::where('reservasi_id','=',$reservasi_id)->update($data);
-
-      ReservasiDetail::where('reservasi_id','=',$reservasi_id)->delete();
-
-      foreach ($list_jadwal as $l) {
-        $detail = new ReservasiDetail();
-        $detail->reservasi_id = $reservasi_id;
-        $detail->jadwal_id = $l;
-        $detail->save();
-        unset($detail);
+        try {
+          Reservasi::where('reservasi_id','=',$reservasi_id)->update($data);
+          ReservasiDetail::where('reservasi_id','=',$reservasi_id)->delete();
+          foreach ($list_jadwal as $l) {
+            $detail = new ReservasiDetail();
+            $detail->reservasi_id = $reservasi_id;
+            $detail->jadwal_id = $l;
+            $detail->save();
+            unset($detail);
+          }
+          $output['code']='1';
+          $output['message'] = 'Edit reservasi berhasil';
+        } catch (Exception $e) {
+          $output['code']='0';
+          $output['message'] = 'Edit reservasi gagal';
+        }
       }
     }
+    return json_encode($output);
   }
 
   function history(Request $request){
     $user_id = $request->input('user_id');
     $jadwal = Jadwal::all()->toArray();
+    $reservasi = array();
     try {
-      $reservasi = Reservasi::where('user_id','=',$user_id)->get()->toArray();
+      $reservasi['reservasi'] = Reservasi::where('user_id','=',$user_id)->get()->toArray();
     } catch (Exception $e) {
       $reservasi['code']=0;
       $reservasi['status']='Request gagal';
       return json_encode((object)$reservasi);
     }
-    if(count($reservasi)==0){
+    if(count($reservasi['reservasi'])==0){
       $reservasi['code']=2;
       $reservasi['status']='Data kosong';
       return json_encode((object)$reservasi);
     } else {
       $j=0;
-      foreach ($reservasi as $r) {
+      foreach ($reservasi['reservasi'] as $r) {
         $detail = ReservasiDetail::where('reservasi_id','=',$r['reservasi_id'])->get()->toArray();
         $len = count($detail);
         $jadwal_id = '';
@@ -121,7 +139,11 @@ class ReservasiController extends Controller
           $jadwal_id .=$jadwal[$d['jadwal_id']-1]['jadwal_start'].'-'.$jadwal[$d['jadwal_id']-1]['jadwal_end'].', ';
           $i++;
         }
-        $reservasi[$j]['jadwal'] = $jadwal_id;
+        $room = Room::where('room_id',$r['room_id'])->first()->toArray();
+        $studio = Studio::where('studio_id',$room['studio_id'])->first()->toArray();
+        $reservasi['reservasi'][$j]['room_nama'] = $room['room_nama'];
+        $reservasi['reservasi'][$j]['studio_nama'] = $studio['studio_nama'];
+        $reservasi['reservasi'][$j]['jadwal'] = $jadwal_id;
         $j++;
       }
     }
@@ -182,5 +204,93 @@ class ReservasiController extends Controller
       $output->status = 'Reservasi tidak ditemukan';
     }
     return json_encode($output);
+  }
+
+  function getKontak(Request $request){
+    $code=array();
+    $reservasi_id = $request->input('reservasi_id');
+    $room_id = Reservasi::where('reservasi_id',$reservasi_id)->first(['room_id'])->toArray()['room_id'];
+    $studio_id = Room::where('room_id',$room_id)->first(['studio_id'])->toArray()['studio_id'];
+    $studio = Studio::where('studio_id',$studio_id)->first(['studio_nama','studio_alamat','studio_telepon','studio_rekening'])->toArray();
+    $output = (object)array();
+    if(count($studio)>0){
+      $output->code = '1';
+      $output->status = 'Request berhasil';
+      $output->studio = $studio;
+      return json_encode($output);
+    } else {
+      $output->code = '0';
+      $output->status = 'Request gagal';
+      return json_encode($output);
+    }
+  }
+
+  function showTransactionPage(){
+    $hak = Session::get('hak');
+    $user_id = Session::get('id');
+    $jadwal = Jadwal::get()->toArray();
+    $city_raw = City::get()->toArray();
+    if($hak == 'ADMIN_ZUPER'){
+      try {
+        $studio_id = Studio::get(['studio_id','studio_nama','city_id'])->toArray();
+        $room_id = Room::whereIn('studio_id',$studio_id)->get(['room_id','room_nama'])->toArray();
+        $reservasi = Reservasi::where('reservasi_status',1)->orWhere('reservasi_status',2)->get();
+      } catch (Exception $e) {
+        $code['c']=0;
+        $code['m']='Request Gagal';
+        Session::flash('msg',$code);
+        return Redirect::back();
+      }
+    } elseif($hak == 'ADMIN_STUDIO'){
+      try {
+        $studio_id = Studio::where('user_id',$user_id)->get(['studio_id','studio_nama','city_id'])->toArray();
+        $room_id = Room::whereIn('studio_id',$studio_id)->get(['room_id','room_nama'])->toArray();
+        $reservasi = Reservasi::whereIn('room_id',$room_id)->where('reservasi_status',1)
+                                 ->orWhere('reservasi_status',2)->get()->toArray();
+      } catch (Exception $e) {
+        $code['c']=0;
+        $code['m']='Request Gagal';  
+        Session::flash('msg',$code);
+        return Redirect::back();
+      }
+    }
+
+    try{
+      //memasukkan room & studio ke array
+      $room = array();
+      $studio = array();
+      $city = array();
+      foreach ($room_id as $r) {
+        $room[$r['room_id']]=$r['room_nama'];
+      }
+      foreach ($studio_id as $s) {
+        $studio[$s['studio_id']]=$s['studio_nama'];
+        $city[$s['studio_id']]=$city_raw[$s['city_id']];
+      }
+      $user_raw = array();
+      $i=0;
+      foreach ($reservasi as $r) {
+        $detail_raw = ReservasiDetail::where('reservasi_id',$r['reservasi_id'])->get()->toArray();
+        $detail = array();
+        foreach ($detail_raw as $d) {
+          $detail[$d['reservasi_detail_id']]=$jadwal[$d['jadwal_id']-1]['jadwal_start'].'-'.$jadwal[$d['jadwal_id']-1]['jadwal_end'];
+        }
+        $reservasi[$i]['detail']=$detail;
+        array_push($user_raw, $r['user_id']);
+        $i++;
+      }
+      $user = User::whereIn('user_id',$user_raw)->get()->toArray();
+      $users = array();
+      foreach ($user as $u) {
+        $users[$u['user_id']]['name']=$u['user_name'];
+        $users[$u['user_id']]['email']=$u['user_email'];
+      }
+    } catch (Exception $e){
+      $code['c']=0;
+      $code['m']='Request Gagal';  
+      Session::flash('msg',$code);
+      return Redirect::back();
+    }
+    return view('transaksi.index')->with(['reservasi'=>$reservasi,'user'=>$users,'room'=>$room,'studio'=>$studio, 'city'=>$city]);
   }
 }
